scraper-microservice/requirements.txt
Flask==3.1.1
asyncpraw==7.8.1
pandas==2.3.1
google-cloud-storage==3.3.0
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-cov==6.2.1
black==25.1.0
ruff==0.12.9